chore: remove definition files as they are no longer needed
